replace tests for edge methods
